+++ lib/modulorails.rb
@@ -7,6 +7,7 @@ require 'generators/modulorails/gitlabci/gitlabci_generator'
 require 'generators/modulorails/healthcheck/health_check_generator'
 require 'generators/modulorails/self_update/self_update_generator'
 require 'generators/modulorails/rubocop/rubocop_generator'
+require 'generators/modulorails/bundleraudit/bundleraudit_generator'
 require 'httparty'
 require 'modulorails/error_data'
 require 'modulorails/success_data'
@@ -163,6 +164,13 @@ module Modulorails
     # Generate a rubocop configuration.
     def generate_rubocop_template
       Modulorails::RubocopGenerator.new([], {}, {}).invoke_all
+    end
+
+    # @author Matthieu 'ciappa_m' Ciappara
+    #
+    # Generate a bundler-audit configuration.
+    def generate_bundleraudit_template
+      Modulorails::BundlerauditGenerator.new([], {}, {}).invoke_all
     end
 
   end